+++ e2e/commands/export.e2e.1.js
@@ -1,9 +1,11 @@
 import path from 'path';
 import fs from 'fs-extra';
-import { expect } from 'chai';
+import chai, { expect } from 'chai';
 import Helper, { VERSION_DELIMITER } from '../../src/e2e-helper/e2e-helper';
 import * as fixtures from '../fixtures/fixtures';
 import { CURRENT_UPSTREAM } from '../../src/constants';
+
+chai.use(require('chai-fs'));
 
 describe('bit export command', function () {
   this.timeout(0);
@@ -674,6 +676,45 @@ describe('bit export command', function () {
           expect(output).to.have.string('was not found');
         });
       });
+      describe('having another staged component without scope-name', () => {
+        let output;
+        let beforeExportScope;
+        before(() => {
+          helper.scopeHelper.getClonedLocalScope(localScopeBefore);
+          helper.scopeHelper.getClonedRemoteScope(remoteScopeBefore);
+          helper.scopeHelper.reInitRemoteScope(anotherRemotePath);
+          helper.fs.outputFile('foo3.js');
+          helper.command.addComponent('foo3.js');
+          helper.command.tagAllComponents();
+          beforeExportScope = helper.scopeHelper.cloneLocalScope();
+        });
+        describe('without defaultScope', () => {
+          before(() => {
+            output = helper.command.export();
+          });
+          it('should not throw an error "toGroupByScopeName() expect ids to have a scope name"', () => {
+            expect(output).to.have.string('exported the following 2 component(s)');
+          });
+          it('should indicate that the component without scope-name was not exported', () => {
+            expect(output).to.have.string('the following component(s) were not exported: foo3');
+          });
+        });
+        describe('with defaultScope', () => {
+          before(() => {
+            helper.scopeHelper.getClonedLocalScope(beforeExportScope);
+            helper.scopeHelper.getClonedRemoteScope(remoteScopeBefore);
+            helper.scopeHelper.reInitRemoteScope(anotherRemotePath);
+            helper.bitJson.addKeyVal(undefined, 'defaultScope', helper.scopes.remote);
+            output = helper.command.export();
+          });
+          it('should export them all successfully', () => {
+            expect(output).to.have.string('exported the following 3 component');
+          });
+          it('should export the non-scope into the defaultScope', () => {
+            expect(output).to.have.string(`${helper.scopes.remote}/foo3`);
+          });
+        });
+      });
     });
     describe('export to a different scope', () => {
       let forkScope;
@@ -845,6 +886,163 @@ describe('bit export command', function () {
           });
         });
       });
+      describe('with --codemod flag', () => {
+        describe('without --set-current-scope', () => {
+          before(() => {
+            helper.scopeHelper.getClonedLocalScope(localScope);
+            helper.scopeHelper.reInitRemoteScope(forkScopePath);
+            helper.fixtures.createComponentUtilsIsString(fixtures.isStringModulePath(helper.scopes.remote));
+            helper.fixtures.createComponentBarFoo(fixtures.barFooModulePath(helper.scopes.remote));
+            helper.env.importDummyCompiler();
+            helper.command.tagScope('1.0.0');
+            helper.command.export(`${forkScope} --include-dependencies --force --codemod`);
+          });
+          it('should not change the files locally on the workspace', () => {
+            const barFoo = helper.fs.readFile('bar/foo.js');
+            expect(barFoo).to.equal(fixtures.barFooModulePath(helper.scopes.remote));
+          });
+          it('should not change the objects locally', () => {
+            const barFoo = helper.command.catComponent(`${helper.scopes.remote}/bar/foo@latest`);
+            const fileHash = barFoo.files[0].file;
+            const fileContent = helper.command.catObject(fileHash);
+            expect(fileContent).to.have.string(helper.scopes.remote);
+            expect(fileContent).to.not.have.string(forkScope);
+          });
+          describe('importing the component from the fork scope to a new workspace', () => {
+            before(() => {
+              helper.scopeHelper.reInitLocalScope();
+              helper.scopeHelper.addRemoteScope(forkScopePath);
+              helper.command.runCmd(`bit import ${forkScope}/bar/foo`);
+            });
+            it('should write the source code with the changed source of the forked scope', () => {
+              const barFoo = helper.fs.readFile('components/bar/foo/foo.js');
+              expect(barFoo).to.have.string(forkScope);
+              expect(barFoo).to.not.have.string(helper.scopes.remote);
+            });
+          });
+        });
+        describe('with --set-current-scope', () => {
+          let localBeforeFork;
+          before(() => {
+            helper.scopeHelper.getClonedLocalScope(localScope);
+            helper.scopeHelper.reInitRemoteScope(forkScopePath);
+            helper.fixtures.createComponentUtilsIsString(fixtures.isStringModulePath(helper.scopes.remote));
+            helper.fixtures.createComponentBarFoo(fixtures.barFooModulePath(helper.scopes.remote));
+            helper.env.importDummyCompiler();
+            helper.command.tagScope('1.0.0');
+            localBeforeFork = helper.scopeHelper.cloneLocalScope();
+            helper.command.export(`${forkScope} --include-dependencies --force --set-current-scope --codemod`);
+          });
+          it('should change the files locally on the workspace', () => {
+            const barFoo = helper.fs.readFile('bar/foo.js');
+            expect(barFoo).to.equal(fixtures.barFooModulePath(forkScope));
+          });
+          // turns out we don't write dists file for author (because author might has its own way of creating dists)
+          it.skip('should change the dist files locally on the workspace', () => {
+            const barFoo = helper.fs.readFile('dist/bar/foo.js');
+            expect(barFoo).to.equal(fixtures.barFooModulePath(forkScope));
+          });
+          it('should change the files objects locally', () => {
+            const barFoo = helper.command.catComponent(`${forkScope}/bar/foo@latest`);
+            const fileHash = barFoo.files[0].file;
+            const fileContent = helper.command.catObject(fileHash);
+            expect(fileContent).to.not.have.string(helper.scopes.remote);
+            expect(fileContent).to.have.string(forkScope);
+          });
+          it('should change the dists objects locally', () => {
+            const barFoo = helper.command.catComponent(`${forkScope}/bar/foo@latest`);
+            const fileHash = barFoo.dists[0].file;
+            const fileContent = helper.command.catObject(fileHash);
+            expect(fileContent).to.not.have.string(helper.scopes.remote);
+            expect(fileContent).to.have.string(forkScope);
+          });
+          it('should be able to require the components and the dependencies', () => {
+            helper.command.build(); // to rewrite the dists. see above 'should change the dist files locally on the workspace'
+            const appJsFixture = `const barFoo = require('@bit/${forkScope}.bar.foo'); console.log(barFoo());`;
+            helper.fs.outputFile('app.js', appJsFixture);
+            const result = helper.command.runCmd('node app.js');
+            expect(result.trim()).to.equal('got is-type and got is-string and got foo');
+          });
+          describe('as imported', () => {
+            let output;
+            before(() => {
+              helper.scopeHelper.getClonedLocalScope(localBeforeFork);
+              helper.command.exportAllComponents();
+
+              helper.scopeHelper.reInitLocalScope();
+              helper.scopeHelper.addRemoteScope();
+              helper.command.importComponent('bar/foo');
+
+              helper.scopeHelper.reInitRemoteScope(forkScopePath);
+              helper.scopeHelper.addRemoteScope(forkScopePath);
+              output = helper.command.export(
+                `${forkScope} --include-dependencies --force --set-current-scope --codemod --all`
+              );
+            });
+            it('should change the files locally on the workspace', () => {
+              const barFoo = helper.fs.readFile('components/bar/foo/foo.js');
+              expect(barFoo).to.equal(fixtures.barFooModulePath(forkScope));
+            });
+            it('should change the dist files locally on the workspace', () => {
+              const barFoo = helper.fs.readFile('components/bar/foo/dist/foo.js');
+              expect(barFoo).to.equal(fixtures.barFooModulePath(forkScope));
+            });
+            it('should change the files objects locally', () => {
+              const barFoo = helper.command.catComponent(`${forkScope}/bar/foo@latest`);
+              const fileHash = barFoo.files[0].file;
+              const fileContent = helper.command.catObject(fileHash);
+              expect(fileContent).to.not.have.string(helper.scopes.remote);
+              expect(fileContent).to.have.string(forkScope);
+            });
+            it('should change the dists objects locally', () => {
+              const barFoo = helper.command.catComponent(`${forkScope}/bar/foo@latest`);
+              const fileHash = barFoo.dists[0].file;
+              const fileContent = helper.command.catObject(fileHash);
+              expect(fileContent).to.not.have.string(helper.scopes.remote);
+              expect(fileContent).to.have.string(forkScope);
+            });
+            it('should not show a warning about untracked files', () => {
+              expect(output).not.to.have.string(
+                'bit did not update the workspace as the component files are not tracked'
+              );
+            });
+            it('should remove the old components from the package.json', () => {
+              const packageJson = helper.packageJson.read();
+              expect(packageJson.dependencies).to.not.have.property(`@bit/${helper.scopes.remote}.bar.foo`);
+              expect(packageJson.dependencies).to.have.property(`@bit/${forkScope}.bar.foo`);
+            });
+            it('should remove the old components from node_modules', () => {
+              const oldComp = `node_modules/@bit/${helper.scopes.remote}.bar.foo`;
+              expect(path.join(helper.scopes.localPath, oldComp)).not.to.be.a.path();
+            });
+            it('should be able to require the components and the dependencies', () => {
+              const appJsFixture = `const barFoo = require('@bit/${forkScope}.bar.foo'); console.log(barFoo());`;
+              helper.fs.outputFile('app.js', appJsFixture);
+              const result = helper.command.runCmd('node app.js');
+              expect(result.trim()).to.equal('got is-type and got is-string and got foo');
+            });
+          });
+        });
+      });
+    });
+  });
+  describe('export with a remote that is not the same as defaultScope', () => {
+    let anotherRemote;
+    let anotherRemotePath;
+    before(() => {
+      helper.scopeHelper.setNewLocalAndRemoteScopes();
+      helper.fixtures.populateWorkspaceWithComponents();
+      helper.bitJson.addKeyVal(undefined, 'defaultScope', helper.scopes.remote);
+      const { scopeName, scopePath } = helper.scopeHelper.getNewBareScope();
+      anotherRemote = scopeName;
+      anotherRemotePath = scopePath;
+      helper.scopeHelper.addRemoteScope(anotherRemotePath);
+      helper.command.tagAllComponents();
+      helper.command.runCmd(`bit export ${anotherRemote} utils/is-type`);
+    });
+    it('should export to the specified remote', () => {
+      const list = helper.command.listScopeParsed(anotherRemote);
+      expect(list).to.have.lengthOf(1);
     });
   });
 });