+++ src/scope/component-ops/export-scope-components.js
@@ -12,7 +12,7 @@ import { ModelComponent, Symlink, Version } from '../models';
 import { getScopeRemotes } from '../scope-remotes';
 import ScopeComponentsImporter from './scope-components-importer';
 import type { Remotes, Remote } from '../../remotes';
-import type Scope from '../scope';
+import Scope from '../scope';
 import { LATEST } from '../../constants';
 import componentIdToPackageName from '../../utils/bit/component-id-to-package-name';
 import Source from '../models/source';
@@ -48,14 +48,25 @@ export async function exportManyBareScope(
   return mergedIds;
 }
 
-export async function exportMany(
+export async function exportMany({
+  scope,
+  ids,
+  remoteName,
+  context = {},
+  includeDependencies = false, // kind of fork. by default dependencies only cached, with this, their scope-name is changed
+  changeLocallyAlthoughRemoteIsDifferent = false, // by default only if remote stays the same the component is changed from staged to exported
+  codemod = false,
+  defaultScope
+}: {
   scope: Scope,
   ids: BitIds,
   remoteName: ?string,
-  context: Object = {},
-  includeDependencies: boolean = false, // kind of fork. by default dependencies only cached, with this, their scope-name is changed
-  changeLocallyAlthoughRemoteIsDifferent: boolean = false // by default only if remote stays the same the component is changed from staged to exported
-): Promise<{ exported: BitIds, updatedLocally: BitIds }> {
+  context?: Object,
+  includeDependencies: boolean,
+  changeLocallyAlthoughRemoteIsDifferent: boolean,
+  codemod: boolean,
+  defaultScope: ?string
+}): Promise<{ exported: BitIds, updatedLocally: BitIds }> {
   logger.debugAndAddBreadCrumb('scope.exportMany', 'ids: {ids}', { ids: ids.toString() });
   enrichContextFromGlobal(context);
   if (includeDependencies) {
@@ -68,7 +79,7 @@ export async function exportMany(
   if (remoteName) {
     return exportIntoRemote(remoteName, ids);
   }
-  const groupedByScope = ids.toGroupByScopeName();
+  const groupedByScope = ids.toGroupByScopeName(defaultScope);
   const results = await pMapSeries(Object.keys(groupedByScope), scopeName =>
     exportIntoRemote(scopeName, groupedByScope[scopeName])
   );
@@ -90,14 +101,15 @@ export async function exportMany(
     const manyObjectsP = componentObjects.map(async (componentObject: ComponentObjects) => {
       const componentAndObject = componentObject.toObjects(scope.objects);
       componentAndObject.component.clearStateData();
-      convertToCorrectScope(scope, componentAndObject, remoteNameStr, includeDependencies, bitIds);
+      await convertToCorrectScope(scope, componentAndObject, remoteNameStr, includeDependencies, bitIds, codemod);
       await changePartialNamesToFullNamesInDists(scope, componentAndObject.component, componentAndObject.objects);
       const remoteObj = { url: remote.host, name: remote.name, date: Date.now().toString() };
       componentAndObject.component.addScopeListItem(remoteObj);
 
-      if (idsToChangeLocally.hasWithoutScope(componentAndObject.component.toBitId())) {
+      if (idsToChangeLocally.hasWithoutScopeAndVersion(componentAndObject.component.toBitId())) {
         componentsAndObjects.push(componentAndObject);
       } else {
+        // the component should not be changed locally. only add the new scope to the scope-list
         const componentAndObjectCloned = componentObject.toObjects(scope.objects);
         componentAndObjectCloned.component.addScopeListItem(remoteObj);
         componentsAndObjects.push(componentAndObjectCloned);
@@ -124,9 +136,12 @@ export async function exportMany(
     await scope.objects.persist();
     // remove version. exported component might have multiple versions exported
     const idsWithRemoteScope: BitId[] = exportedIds.map(id => BitId.parse(id, true).changeVersion(null));
+    const idsWithRemoteScopeUniq = BitIds.uniqFromArray(idsWithRemoteScope);
     return {
-      exported: BitIds.uniqFromArray(idsWithRemoteScope),
-      updatedLocally: BitIds.uniqFromArray(idsWithRemoteScope.filter(id => idsToChangeLocally.hasWithoutScope(id)))
+      exported: idsWithRemoteScopeUniq,
+      updatedLocally: BitIds.fromArray(
+        idsWithRemoteScopeUniq.filter(id => idsToChangeLocally.hasWithoutScopeAndVersion(id))
+      )
     };
   }
 
@@ -180,45 +195,22 @@ async function mergeObjects(scope: Scope, manyObjects: ComponentTree[]): Promise
  * to the bare-scope name.
  * Since the changes it does affect the Version objects, the version REF of a component, needs to be changed as well.
  */
-function convertToCorrectScope(
+async function convertToCorrectScope(
   scope: Scope,
   componentsObjects: { component: ModelComponent, objects: BitObject[] },
   remoteScope: string,
   fork: boolean,
-  exportingIds: BitIds
-): void {
-  const getIdWithUpdatedScope = (dependencyId: BitId): BitId => {
-    if (dependencyId.scope === remoteScope) {
-      return dependencyId; // nothing has changed
-    }
-    if (!dependencyId.scope || fork || exportingIds.hasWithoutVersion(dependencyId)) {
-      const depId = ModelComponent.fromBitId(dependencyId);
-      // todo: use 'load' for async and switch the foreach with map.
-      const dependencyObject = scope.objects.loadSync(depId.hash());
-      if (dependencyObject instanceof Symlink) {
-        return dependencyId.changeScope(dependencyObject.realScope);
-      }
-      return dependencyId.changeScope(remoteScope);
-    }
-    return dependencyId;
-  };
-
-  const getBitIdsWithUpdatedScope = (bitIds: BitIds): BitIds => {
-    const updatedIds = bitIds.map(id => getIdWithUpdatedScope(id));
-    return BitIds.fromArray(updatedIds);
-  };
-
-  componentsObjects.objects.forEach((object: BitObject) => {
-    if (object instanceof Version) {
-      const hashBefore = object.hash().toString();
-      object.getAllDependencies().forEach((dependency) => {
-        dependency.id = getIdWithUpdatedScope(dependency.id);
-      });
-      object.flattenedDependencies = getBitIdsWithUpdatedScope(object.flattenedDependencies);
-      object.flattenedDevDependencies = getBitIdsWithUpdatedScope(object.flattenedDevDependencies);
-      object.flattenedCompilerDependencies = getBitIdsWithUpdatedScope(object.flattenedCompilerDependencies);
-      object.flattenedTesterDependencies = getBitIdsWithUpdatedScope(object.flattenedTesterDependencies);
-      const hashAfter = object.hash().toString();
+  exportingIds: BitIds,
+  codemod: boolean
+): Promise<void> {
+  // $FlowFixMe
+  const versionsObjects: Version[] = componentsObjects.objects.filter(object => object instanceof Version);
+  await Promise.all(
+    versionsObjects.map(async (objectVersion: Version) => {
+      const hashBefore = objectVersion.hash().toString();
+      if (codemod) await _replaceSrcOfVersionIfNeeded(objectVersion);
+      changeDependencyScope(objectVersion);
+      const hashAfter = objectVersion.hash().toString();
       if (hashBefore !== hashAfter) {
         logger.debugAndAddBreadCrumb(
           'scope._convertToCorrectScope',
@@ -232,10 +224,87 @@ function convertToCorrectScope(
           }
         });
       }
-    }
-  });
-
+    })
+  );
   componentsObjects.component.scope = remoteScope;
+
+  function changeDependencyScope(version: Version): void {
+    version.getAllDependencies().forEach((dependency) => {
+      dependency.id = getIdWithUpdatedScope(dependency.id);
+    });
+    version.flattenedDependencies = getBitIdsWithUpdatedScope(version.flattenedDependencies);
+    version.flattenedDevDependencies = getBitIdsWithUpdatedScope(version.flattenedDevDependencies);
+    version.flattenedCompilerDependencies = getBitIdsWithUpdatedScope(version.flattenedCompilerDependencies);
+    version.flattenedTesterDependencies = getBitIdsWithUpdatedScope(version.flattenedTesterDependencies);
+  }
+
+  function getIdWithUpdatedScope(dependencyId: BitId): BitId {
+    if (dependencyId.scope === remoteScope) {
+      return dependencyId; // nothing has changed
+    }
+    if (!dependencyId.scope || fork || exportingIds.hasWithoutVersion(dependencyId)) {
+      const depId = ModelComponent.fromBitId(dependencyId);
+      // todo: use 'load' for async and switch the foreach with map.
+      const dependencyObject = scope.objects.loadSync(depId.hash());
+      if (dependencyObject instanceof Symlink) {
+        return dependencyId.changeScope(dependencyObject.realScope);
+      }
+      return dependencyId.changeScope(remoteScope);
+    }
+    return dependencyId;
+  }
+  function getBitIdsWithUpdatedScope(bitIds: BitIds): BitIds {
+    const updatedIds = bitIds.map(id => getIdWithUpdatedScope(id));
+    return BitIds.fromArray(updatedIds);
+  }
+  async function _replaceSrcOfVersionIfNeeded(version: Version) {
+    const files = [...version.files, ...(version.dists || [])];
+    await Promise.all(
+      files.map(async (file) => {
+        const newFileObject = await _createNewFileIfNeeded(version, file);
+        if (newFileObject) {
+          file.file = newFileObject.hash();
+          componentsObjects.objects.push(newFileObject);
+        }
+        return null;
+      })
+    );
+  }
+  async function _createNewFileIfNeeded(version: Version, file: Object): Promise<?Source> {
+    const currentHash = file.file;
+    // $FlowFixMe
+    const fileObject: Source = await scope.objects.load(currentHash);
+    const fileString = fileObject.contents.toString();
+    const dependenciesIds = version.getAllDependencies().map(d => d.id);
+    const allIds = [...dependenciesIds, componentsObjects.component.toBitId()];
+    let newFileString = fileString;
+    allIds.forEach((id) => {
+      if (id.scope === remoteScope) {
+        return; // nothing to do, the remote has not changed
+      }
+      const idWithNewScope = id.changeScope(remoteScope);
+      const pkgNameWithNewScope = componentIdToPackageName(idWithNewScope, componentsObjects.component.bindingPrefix);
+      const pkgNameWithOldScope = componentIdToPackageName(id, componentsObjects.component.bindingPrefix);
+      const singleQuote = "'";
+      const doubleQuotes = '"';
+      [singleQuote, doubleQuotes].forEach((quoteType) => {
+        // replace an exact match. (e.g. '@bit/old-scope.is-string' => '@bit/new-scope.is-string')
+        newFileString = newFileString.replace(
+          new RegExp(quoteType + pkgNameWithOldScope + quoteType, 'g'),
+          quoteType + pkgNameWithNewScope + quoteType
+        );
+        // the require/import statement might be to an internal path (e.g. '@bit/david.utils/is-string/internal-file')
+        newFileString = newFileString.replace(
+          new RegExp(`${quoteType}${pkgNameWithOldScope}/`, 'g'),
+          `${quoteType}${pkgNameWithNewScope}/`
+        );
+      });
+    });
+    if (newFileString !== fileString) {
+      return Source.from(Buffer.from(newFileString));
+    }
+    return null;
+  }
 }
 
 /**
@@ -270,8 +339,13 @@ async function changePartialNamesToFullNamesInDists(
 
   async function _createNewDistIfNeeded(version: Version, dist: Object): Promise<?Source> {
     const currentHash = dist.file;
+    // if a dist file has changed as a result of codemod, it's not on the fs yet, so we fallback
+    // to load from the objects it was pushed before. it'd be better to have more efficient mechanism.
+    // currently, we require calculating the hash for each one of the source every time.
     // $FlowFixMe
-    const distObject: Source = await scope.objects.load(currentHash);
+    const distObject: Source =
+      (await currentHash.load(scope.objects)) ||
+      objects.filter(obj => obj instanceof Source).find(obj => obj.hash().toString() === currentHash.toString());
     const distString = distObject.contents.toString();
     const dependenciesIds = version.getAllDependencies().map(d => d.id);
     const allIds = [...dependenciesIds, component.toBitId()];