+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- [#1956](https://github.com/teambit/bit/issues/1956) introduce a new flag `--codemod` for `bit export` to replace the import/require statements in the source to the newly exported scope
+- fix "Converting circular structure to JSON" error when logging a circular metadata object
+- fix exporting to a different scope than workspace configuration of `defaultScope`
+- fix exporting components with and without scope at the same time
+- [#1999](https://github.com/teambit/bit/issues/1999) show a descriptive error when a component is missing from the scope
 - block tagging components with prerelease versions
 - [#1981](https://github.com/teambit/bit/issues/1981) allow compilers to add all dependencies types and not only devDependencies